completely redesigned
most of keys are positioned to a similar position on the HHKB layout,
although they might be different layers.

layer 2 is the layer of mouse navigation. left/up movements are on the
left-hand side. right/down movements are on the right-hand side.

layer 1 is the layer of digits and punctuations.

layer 0 is the layer of letters and for most of editing. could still
require a bit more optimization.